Added my FP Choice Information
Jack, Conor, and Caden plan intend to make an algorithm that takes a CSV file or stock prices over a given period of time and returns a portfolio of stocks that have the highest mean real return and lowest standard deviation from that file. We intend to make it take any number of stocks, calculate the mean real growth per period and the observed standard deviation over that period, and create a bundle that will maximize growth given an ideal standard deviation.